ref(enchantments): Migrate enhanced vanilla enchantments to abstract class
